# docs/swagger.yaml
basePath: /v1
definitions:
  main.CreateTransactionRequest:
    properties:
      amount:
        type: number
      categoryName:
        type: string
      description:
        type: string
      transactionDate:
        description: Assuming this is a string for simplicity, could be time.Time
        type: string
      transactionType:
        type: string
    type: object
...
  main.UpdateTransactionRequest:
    properties:
      amount:
        type: number
      categoryName:
        type: string
      description:
        type: string
      transactionDate:
        description: Assuming this is a string for simplicity, could be time.Time
        type: string
      transactionType:
        type: string
    type: object
...
  store.Transaction:
    properties:
      amount:
        type: number
      categoryId:
        type: integer
      categoryName:
        type: string
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      transactionDate:
        description: Assuming this is a string for simplicity, could be time.Time
        type: string
      transactionType:
        type: string
      updatedAt:
        type: string
      userId:
        type: integer
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for a finance tracker application
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: FinTracker API
paths:
...
  /transactions:
    get:
      description: List transactions for the authenticated user
      parameters:
      - description: Start date in RFC3339 format
        in: query
        name: startDate
        type: string
      - description: End date in RFC3339 format
        in: query
        name: endDate
        type: string
      - description: Transaction type (income/expense)
        in: query
        name: transactionType
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.Transaction'
            type: array
        "401":
          description: Unauthorized
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: List transactions for the authenticated user
      tags:
      - transactions
    post:
      consumes:
      - application/json
      description: Create a new transaction for the authenticated user
      parameters:
      - description: Transaction data
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/main.CreateTransactionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/store.Transaction'
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Create a new transaction
      tags:
      - transactions
  /transactions/{id}:
    delete:
      description: Delete a transaction by ID for the authenticated user
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Delete a transaction by ID
      tags:
      - transactions
    get:
      description: Get a transaction by ID for the authenticated user
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Transaction'
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Get a transaction by ID
      tags:
      - transactions
    patch:
      description: Update a transaction by ID for the authenticated user
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: integer
      - description: Transaction data
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/main.UpdateTransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Transaction'
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Update a transaction by ID
      tags:
      - transactions
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"